+++ src/net_prof/__init__.py
@@ -1,16 +1,17 @@
 # src/net_prof/__init__.py
 
-from .engine import summarize, collect, dump, dump_html, collect
-from .visualize import bar_chart, heat_map, generate_iface_barchart
+from .engine import summarize, collect, dump, dump_html
+from .visualize import generate_iface_barchart, non_zero_bar_chart, heat_map, unit_barchart, group_barchart
 
 __all__ = [
     "summarize",
     "collect",
     "dump",
     "dump_html",
-    "collect",
-    "bar_chart",
+    "generate_iface_barchart",
+    "non_zero_bar_chart",
     "heat_map",
-    "generate_iface_barchart"
+    "unit_barchart",
+    "group_barchart"
 ]